Extract DeepSeekV2 async CAM forward flow
